requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-location-field==2.1.0
gunicorn==20.1.0
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
Unipath==1.1
whitenoise==6.2.0
